fix(rust): close router channel after run
Since messages are no longer processed when `Router::run` finishes and
senders may wait to send more we close the receiver to unblock them.

NB that there is no guarantee that all sent messages are actually
processed by the router as there may be unprocessed messages lingering in
the channel queue. This behaviour is not changed by this PR.